condaenv.fxqsesx2.requirements.txt
absl-py==1.4.0
accelerate==0.22.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
av==10.0.0
babel==2.12.1
backcall==0.2.0
bddl==3.5.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.2
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
cython==0.29.37
git+https://github.com/Farama-Foundation/d4rl@71a9549f2091accff93eeff68f1f3ab2c0e0a288#egg=d4rl
debugpy==1.6.7.post1
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
dm-control==1.0.14
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
easydict==1.13
einops==0.6.1
evdev==1.6.1
exceptiongroup==1.1.3
executing==1.2.0
fasteners==0.18
fastjsonschema==2.18.0
filelock==3.12.2
fonttools==4.42.1
future==0.18.3
gitdb==4.0.10
gitpython==3.1.32
glfw==2.6.2
gym==0.23.1
gdown==5.1.0
h5py==3.9.0
huggingface-hub==0.22.2
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-resources==6.0.1
iopath==0.1.10
ipdb==0.13.13
ipykernel==6.25.1
ipython==8.12.2
ipywidgets==8.1.0
jedi==0.19.0
jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-client==8.3.0
jupyter-core==5.3.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter-server==2.7.2
jupyter-server-terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-server==2.24.0
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
labmaze==1.0.6
lit==16.0.6
lxml==4.9.3
markupsafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.5
mujoco==2.3.7
mujoco-py==2.1.2.14
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook-shim==0.2.3
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
opencv-python==4.8.0.76
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
patchelf==0.17.2.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.0.0
pkgutil-resolve-name==1.3.10
platformdirs==3.10.0
prettytable==3.8.0
proglog==0.1.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pybullet==3.2.5
pycparser==2.21
pygame==2.5.2
pygments==2.16.1
pymunk==6.6.0
pynput==1.7.6
pynvml==11.5.0
pyopengl==3.1.7
pyopengl-accelerate==3.1.7
pyparsing==3.0.9
python-json-logger==2.0.7
python-xlib==0.33
pyyaml==6.0.1
pyzmq==25.1.0
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
robosuite==1.4.1
rpds-py==0.9.2
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.10.1
send2trash==1.8.2
sentry-sdk==1.29.2
setproctitle==1.3.2
shapely==2.0.3
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
submitit==1.5.1
sympy==1.12
tables==3.8.0
tabulate==0.9.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.16
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
triton==2.0.0
typing-extensions==4.7.1
urllib3==2.0.4
wandb==0.15.8
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.8
zarr==2.16.1
zipp==3.16.2
git+https://github.com/Farama-Foundation/d4rl.git